requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
finnhub-python==2.4.19
requests==2.31.0
yfinance==0.2.66
praw==7.8.1
pandas==2.3.3
numpy==2.3.4
pytz==2025.2
ta==0.11.0

